+++ api/kogito-api/src/main/java/org/kie/kogito/event/CloudEventEmitter.java
@@ -12,10 +12,9 @@
  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
- *
  */
 
-package org.kie.kogito.services.event;
+package org.kie.kogito.event;
 
 import java.util.concurrent.CompletionStage;
