+++ kogito-codegen/src/main/resources/class-templates/SpringMessageProducerTemplate.java
@@ -2,12 +2,9 @@ package org.kie.kogito.test;
 
 import org.kie.api.runtime.process.ProcessInstance;
 import org.kie.kogito.conf.ConfigBean;
-import org.kie.kogito.services.event.CloudEventEmitter;
-import org.kie.kogito.services.event.EventMarshaller;
+import org.kie.kogito.event.CloudEventEmitter;
 import org.kie.kogito.event.impl.DefaultEventMarshaller;
 import org.kie.kogito.services.event.impl.AbstractMessageProducer;
-
-import java.util.Optional;
 
 @org.springframework.stereotype.Component()
 public class MessageProducer extends AbstractMessageProducer<$DataType$, $DataEventType$> {
